Made some changes in preparation for revamping the README with new screenshots. Changes related to the Contact Us Form: Changed the Donate nav link to a Contact Us nav link which scrolls to the Contact Us form at /about-us#contact-us-section. Integrated web3forms and react-hook-form to the Contact Us form so that when the user submits a form, an email is sent to my email (sara.marie.hines1@gmail.com) with the user's submitted formData and message. Also implemented additional input validation and protected the web3forms API key by setting up a GET route to /api/contact-us in server.js to provide the frontend with the API Key from the backend. Added success (or error) message upon form submission with a fade-out effect. The Contact Us form is now its own component, ContactUs, rendered inside of AboutUs.jsx. Added ContactUs.css and ContactUs.scss for additional styling.
In AboutUs.css, MythIndex.css, StoryIndex.css, and UserProfile.css, made some changes for improved responsiveness and mobile styling. In CreateStory.css, fixed a few styling inconsistencies in the modals after the first modal. In Header.css, made some changes to the styling of the login and signup modals for focus effects, spacing of input validation messages, and centered 'x's for the close buttons. Added additional comments in files across the repo.